Move EmptyCollectionFilter to core
- Renamed to EmptyConfigCollectionFilter
- With the intent of re-use across projects
